Sharing module

+++ SetlistManager.Api/Controllers/UsersController.cs
@@ -7,32 +7,23 @@ using SetlistManager.Business.Services;
 using SetlistManager.Common.Genius.Models;
 using SetlistManager.Common.Models;
 
-namespace SetlistManager. Api.Controllers;
+namespace SetlistManager.Api.Controllers;
 
 [Route("api/users")]
 public class UsersController : BaseController
 {
     private readonly IUserService _userService;
-    private readonly ISetlistsService _setlistsService;
     private readonly ICurrentUserContext _currentUserContext;
     private readonly IGeniusAuthService _geniusAuthService;
     private readonly IOptions<AppOptions> _appOptions;
 
-    public UsersController(IUserService userService, ISetlistsService setlistsService, ICurrentUserContext currentUserContext,
+    public UsersController(IUserService userService, ICurrentUserContext currentUserContext, 
         IGeniusAuthService geniusAuthService, IOptions<AppOptions> appOptions)
     {
         _appOptions = appOptions;
         _userService = userService;
-        _setlistsService = setlistsService;
         _currentUserContext = currentUserContext;
         _geniusAuthService = geniusAuthService;
-    }
-
-    [HttpPut]
-    public async Task<ActionResult> UpdateUser(UserModel model)
-    {
-        await _userService.UpdateUserAsync(model);
-        return NoContent();
     }
 
     [HttpGet]
@@ -44,18 +35,17 @@ public class UsersController : BaseController
     [HttpGet("{id}")]
     public async Task<ActionResult<UserModel>> GetUserById(int id)
     {
-        var userId = _currentUserContext.GetCurrentUserId();
-
-        if (userId is null)
+        if (!IsAuthorized(id))
             return Unauthorized();
 
-        return Ok(await _userService.GetCurrentUserAsync(userId.Value));
+        return Ok(await _userService.GetCurrentUserAsync(id));
     }
 
-    [HttpGet("{id}/setlists")]
-    public async Task<ActionResult<List<SetlistModel>>> GetUserSetlists(int id)
+    [HttpPut]
+    public async Task<ActionResult> UpdateUser(UserModel model)
     {
-        return Ok(await _setlistsService.GetAllSetlistsOfUserAsync(id));
+        await _userService.UpdateUserAsync(model);
+        return NoContent();
     }
 
     [AllowAnonymous]
@@ -88,11 +78,14 @@ public class UsersController : BaseController
             return BadRequest("Could not add token to user, provider not found");
 
         return Redirect(_appOptions.Value.UserPortalUrl);
-    }
+    }    
 
     [HttpPost("{id}/friendships")]
     public async Task<ActionResult> InitiateFriendship(int id, [FromBody] FriendshipRequestModel requestModel)
     {
+        if (!IsAuthorized(id))
+            return Unauthorized();
+
         await _userService.HandleFriendshipRequestAsync(id, requestModel);
         
         return NoContent();
@@ -101,6 +94,9 @@ public class UsersController : BaseController
     [HttpDelete("{id}/friendships/{friendshipId}")]
     public async Task<ActionResult> RemoveFriendship(int id, int friendshipId)
     {
+        if (!IsAuthorized(id))
+            return Unauthorized();
+
         await _userService.RemoveFriendshipAsync(id, friendshipId);
         
         return NoContent();
@@ -108,10 +104,8 @@ public class UsersController : BaseController
 
     [HttpGet("{id}/friendships")]
     public async Task<ActionResult<PagedResponse<FriendModel>>> GetUserFriends(int id, [FromQuery] PagedRequest pagedRequest)
-    {
-        var currentUserId = _currentUserContext.GetCurrentUserId();
-        
-        if (currentUserId is null)
+    {        
+        if (!IsAuthorized(id))
             return Unauthorized();
         
         var result = await _userService.GetUserFriendsAsync(id, pagedRequest);
@@ -121,9 +115,18 @@ public class UsersController : BaseController
 
     [HttpPut("{id}/friendships/{friendshipId}")]
     public async Task<ActionResult> AcceptFriendship(int id, int friendshipId)
-    {        
+    {
+        if (!IsAuthorized(id))
+            return Unauthorized();
+
         await _userService.AcceptFriendshipAsync(id, friendshipId);
         
         return NoContent();
+    }
+
+    private bool IsAuthorized(int id)
+    {
+        var currentUserId = _currentUserContext.GetCurrentUserId();
+        return currentUserId is not null && currentUserId.Value == id;
     }
 }
